requirements.txt
Flask==0.12.1
Jinja2==2.9.6
MarkupSafe==1.0
RPi.GPIO==0.6.3
Werkzeug==0.12.1
click==6.7
itsdangerous==0.24
pyserial==3.3
